+++ dbxcli_sync/cli.py
@@ -1,4 +1,5 @@
 from .sync import DbxcliSync
+from .getr import getr
 import click
 
 @click.group()
@@ -11,12 +12,30 @@ def cli():
 @click.argument('dbxdir', type=str)
 @click.option('--verbosity', default=0, help="Verbosity level: 0, 1, 2")
 def sync(localdir, dbxdir, verbosity):
-    """dbxcli_sync command that solves https://github.com/dropbox/dbxcli/issues/53"""
+    """
+    sync local folder with remote dropbox folder
+    Solves https://github.com/dropbox/dbxcli/issues/53
+    """
     dcs = DbxcliSync(localdir, dbxdir, verbosity)
     dcs.sync_dir()
 
 
+@click.command()
+@click.argument('dbxdir', type=str)
+@click.argument('localdir', type=click.Path(exists=True))
+@click.option('--verify/--no-verify', default=False, help="Verify downloads (download it several times)")
+@click.option('--verbose/--no-verbose', default=False, help="Be verbose")
+def getr(dbxdir, localdir, verify, verbose):
+    """
+    Recursive get
+    Solves https://github.com/dropbox/dbxcli/issues/60
+    """
+    getr(dbxdir, localdir, verify, verbose)
+
+
+
 cli.add_command(sync)
+cli.add_command(getr)
 
 
 #if __name__ == '__main__':