+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.0.2 (2021-07-27)
 
 - factor out the sync command and make it a subcommand in preparation to add recursive get
+- add getr function for recursive get from https://gist.github.com/debuti/5887c126811eeae1bf9451e73a7b8fd8
 
 
 ## v0.0.1 (2021-07-23)